Updated the buffered log to print in openRealmBefore hook
